Moved SerializerTestBase to s-core and resolved Maven dependencies
