+++ stratosphere-core/src/test/java/eu/stratosphere/api/common/typeutils/SerializerTestBase.java
@@ -12,7 +12,7 @@
  * specific language governing permissions and limitations under the License.
  *
  **********************************************************************************************************************/
-package eu.stratosphere.api.java.typeutils.runtime;
+package eu.stratosphere.api.common.typeutils;
 
 import static org.junit.Assert.*;
