+++ stratosphere-java/pom.xml
@@ -23,6 +23,13 @@
 			<version>${project.version}</version>
 		</dependency>
 		<dependency>
+			<groupId>eu.stratosphere</groupId>
+			<artifactId>stratosphere-core</artifactId>
+			<version>${project.version}</version>
+			<type>test-jar</type>
+  			<scope>test</scope>
+		</dependency>
+		<dependency>
 			<groupId>org.apache.avro</groupId>
 			<artifactId>avro</artifactId>
 			<version>1.7.5</version>
