+++ stratosphere-java/src/test/java/eu/stratosphere/api/java/typeutils/runtime/LongArraySerializerTest.java
@@ -14,6 +14,7 @@
  **********************************************************************************************************************/
 package eu.stratosphere.api.java.typeutils.runtime;
 
+import eu.stratosphere.api.common.typeutils.SerializerTestBase;
 import eu.stratosphere.api.common.typeutils.TypeSerializer;
 
 /**
